+++ backend/src/main/java/COMP_49X_our_search/backend/database/repositories/MajorRepository.java
@@ -3,4 +3,4 @@ package COMP_49X_our_search.backend.database.repositories;
 import COMP_49X_our_search.backend.database.entities.Major;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MajorRepository extends JpaRepository<Major, Long> {}
+public interface MajorRepository extends JpaRepository<Major, Integer> {}
