+++ backend/src/main/java/COMP_49X_our_search/backend/database/repositories/ResearchPeriodRepository.java
@@ -3,4 +3,4 @@ package COMP_49X_our_search.backend.database.repositories;
 import COMP_49X_our_search.backend.database.entities.ResearchPeriod;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ResearchPeriodRepository extends JpaRepository<ResearchPeriod, Long> {}
+public interface ResearchPeriodRepository extends JpaRepository<ResearchPeriod, Integer> {}
