+++ backend/src/main/java/COMP_49X_our_search/backend/database/repositories/StudentRepository.java
@@ -3,4 +3,4 @@ package COMP_49X_our_search.backend.database.repositories;
 import COMP_49X_our_search.backend.database.entities.Student;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface StudentRepository extends JpaRepository<Student, Long> {}
+public interface StudentRepository extends JpaRepository<Student, Integer> {}
