+++ backend/src/main/java/COMP_49X_our_search/backend/database/repositories/UmbrellaTopicRepository.java
@@ -3,4 +3,4 @@ package COMP_49X_our_search.backend.database.repositories;
 import COMP_49X_our_search.backend.database.entities.UmbrellaTopic;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UmbrellaTopicRepository extends JpaRepository<UmbrellaTopic, Long> {}
+public interface UmbrellaTopicRepository extends JpaRepository<UmbrellaTopic, Integer> {}
